Added authorization to comment APIs

+++ src/main/java/com/bhargav/blog/blog_api/controller/CommentController.java
@@ -2,8 +2,10 @@ package com.bhargav.blog.blog_api.controller;
 
 import com.bhargav.blog.blog_api.model.Comment;
 import com.bhargav.blog.blog_api.service.CommentService;
-import lombok.AllArgsConstructor;
+import jakarta.servlet.http.HttpServletRequest;
+import com.bhargav.blog.blog_api.config.JwtUtil;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,22 +16,36 @@ public class CommentController {
 
     private final CommentService commentService;
 
+    private final JwtUtil jwtUtil;
+
     @PostMapping("/{postId}/addComment")
-    public ResponseEntity<String> createComment(@PathVariable int postId, @RequestBody Comment comment) {
+    public ResponseEntity<String> createComment(@PathVariable int postId, @RequestBody Comment comment, HttpServletRequest request) {
+
+        if(!jwtUtil.isAuthorized(request)){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
 
         return commentService.createComment(postId, comment);
 
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<String> updateComment(@PathVariable int id, @RequestBody Comment comment) {
+    public ResponseEntity<String> updateComment(@PathVariable int id, @RequestBody Comment comment, HttpServletRequest request) {
+
+        if(!jwtUtil.isAuthorized(request)){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
 
         return commentService.updateComment(id, comment);
 
     }
 
     @PutMapping("/delete/{id}")
-    public ResponseEntity<String> deleteComment(@PathVariable int id) {
+    public ResponseEntity<String> deleteComment(@PathVariable int id, HttpServletRequest request) {
+
+        if(!jwtUtil.isAuthorized(request)){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
 
         return commentService.deleteComment(id);
